# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API สำหรับ blog — อ่านได้ทุกคน แก้ได้เฉพาะเจ้าของ",
        "title": "Blog User API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "list user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "หน้า เริ่มที่ 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "ต่อหน้า สูงสุด 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "ดู user ตาม id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "id ไม่ใช่ UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_response.Body": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Pagination"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ใส่ \"Bearer \" นำหน้า token ที่ได้จาก /api/v1/auth/login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}